refactor(net): extract NAT engine into arcbox-conntrack crate (Gate A)
Move the connection-tracking table, zero-copy packet translation (SNAT/reverse-NAT), NatEngineConfig, and the pure HostNetIO egress trait out of arcbox-net into a new arcbox-conntrack crate, depending only on arcbox-packet + arcbox-datapath + hashbrown. arcbox-net re-exports it via a thin nat_engine.rs, preserving arcbox_net::nat_engine::* (including ::checksum and ::conntrack). conntrack.rs now takes CachePadded from arcbox-datapath, the DarwinTun HostNetIO impl points at crate::nat_engine, and hashbrown drops out of arcbox-net (only conntrack used it).

Completes Gate A: arcbox-packet, arcbox-datapath, and arcbox-conntrack all compile standalone with no dependency on arcbox-net/arcbox-virtio/arcbox-vmm/arcbox-vz (verified via cargo tree). arcbox-conntrack: 12 tests, clippy -D warnings clean; arcbox-net clippy and test suites green (0 failed); arcbox-vmm consumer checks clean.
