+++ virt/arcbox-net/Cargo.toml
@@ -9,6 +9,7 @@ repository.workspace = true
 authors.workspace = true
 
 [dependencies]
+arcbox-conntrack = { workspace = true }
 arcbox-datapath = { workspace = true }
 arcbox-dhcp = { workspace = true }
 arcbox-error = { workspace = true }
@@ -23,7 +24,6 @@ libc = { workspace = true }
 chrono = { workspace = true }
 uuid = { workspace = true }
 ipnetwork = "0.20"
-hashbrown = "0.14"
 crossbeam-utils = "0.8"
 socket2 = { version = "0.5", features = ["all"] }
 tokio-util = "0.7"
