+++ virt/arcbox-net/src/nat_engine.rs
@@ -1,0 +1,5 @@
+//! NAT engine, re-exported from the extracted `arcbox-conntrack` crate.
+//! Preserves the `arcbox_net::nat_engine::*` paths (including `::checksum`
+//! and `::conntrack`) for existing callers.
+
+pub use arcbox_conntrack::*;
